requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
flask~=2.3.3
click
itsdangerous
MarkupSafe
gunicorn
Werkzeug